Match fully functional + Bug Fixes
Match functional - Multiple users can join in a lobby, go into match and shows everyone's cursors and ripples. Tested for users leaving both lobby and match with checks to make sure the host is reassigned if host left.
- In order to send to everyone in a match, I had to change the Send signature for BCRelayProxy for "toNetId" from an int to int64 in order to pass the value needed to send to all players.
- Users can leave the match and return to logged in menu to then search for another lobby.

+++ RelayTestApp/Plugins/BCClient/Source/BCClientPlugin/Private/BlueprintProxies/BCRelayProxy.cpp
@@ -74,7 +74,7 @@ void UBCRelayProxy::DeregisterSystemCallback(UBrainCloudWrapper *brainCloudWrapp
 	UBCWrapperProxy::GetBrainCloudInstance(brainCloudWrapper)->getRelayService()->deregisterSystemCallback();
 }
 
-void UBCRelayProxy::Send(UBrainCloudWrapper *brainCloudWrapper, const TArray<uint8> &data, int toNetId, bool reliable, bool ordered, BCRelayChannel channel)
+void UBCRelayProxy::Send(UBrainCloudWrapper *brainCloudWrapper, const TArray<uint8> &data, int64 toNetId, bool reliable, bool ordered, BCRelayChannel channel)
 {
 	uint64 to_netId = (uint64)toNetId;
 	if (toNetId == -1)
